worker/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-pubsub
pydantic
